+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.8.5
+
+- --query option can be used multiple times.
+- Replace structopt crate with clap.
+
 ## v0.8.4
 
 - Report an error when team members are needed but `read:org` scope is not in the GitHub token to get them.